+++ react/legacy/components/ProductSummarySmall.js
@@ -23,6 +23,7 @@ class ProductSummarySmall extends Component {
       nameProps,
       priceProps,
       buyButtonProps,
+      inViewRef,
     } = this.props
 
     const containerClasses = classNames(
@@ -60,6 +61,7 @@ class ProductSummarySmall extends Component {
         className={containerClasses}
         onMouseEnter={handleMouseEnter}
         onMouseLeave={handleMouseLeave}
+        ref={inViewRef}
       >
         <Link
           className={`${styles.clearLink} flex flex-column`}