+++ CHANGELOG.md
@@ -6,9 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Product Impression Events on legacy product summary.
 
 ## [2.53.1] - 2020-04-13
-
 ### Fixed
 - Product Summary Image - README.md Documentation